chore: rename interfaces path
