--- src/transaction/interface/index.ts
@@ -1,1 +1,0 @@
-export * from './transaction-repository';
