+++ src/transaction/transaction.module.ts
@@ -5,10 +5,11 @@ import { TransactionController } from '@/shared/infra/http/controllers/';
 import { Transaction } from '@/transaction/infra/entities';
 import { Wallet } from '@/wallet/infra/entities';
 import { TransactionRepository } from './infra/repositories';
-import { ITransactionRepository } from './interface';
+import { ITransactionRepository } from './interfaces';
 import {
   DeleteTransactionUseCase,
   RegisterTransactionUseCase,
+  UpdateTransactionUseCase,
 } from '@/transaction/use-cases';
 import { WalletRepository } from '@/wallet/infra/repositories';
 import { IWalletRepository } from '@/wallet/interfaces';
@@ -19,6 +20,7 @@ import { IWalletRepository } from '@/wallet/interfaces';
   providers: [
     RegisterTransactionUseCase,
     DeleteTransactionUseCase,
+    UpdateTransactionUseCase,
     {
       provide: ITransactionRepository,
       useClass: TransactionRepository,
