+++ src/transaction/use-cases/delete/delete.spec.ts
@@ -2,7 +2,7 @@ import { NotFoundException } from '@nestjs/common';
 import { Test, TestingModule } from '@nestjs/testing';
 
 import { Transaction } from '@/transaction/infra/entities';
-import { ITransactionRepository } from '@/transaction/interface';
+import { ITransactionRepository } from '@/transaction/interfaces';
 import { DeleteTransactionUseCase } from '@/transaction/use-cases';
 
 describe('Delete Transaction Use Case', () => {
