+++ src/transaction/use-cases/register/register.ts
@@ -3,7 +3,7 @@ import { Injectable, NotFoundException } from '@nestjs/common';
 import { TransactionType } from '@/shared/constants';
 import { CreateTransactionDTO } from '@/transaction/dto';
 import { Transaction } from '@/transaction/infra/entities';
-import { ITransactionRepository } from '@/transaction/interface';
+import { ITransactionRepository } from '@/transaction/interfaces';
 import { IWalletRepository } from '@/wallet/interfaces';
 
 @Injectable()
